DW-762 merge linbit patch [1d45dcb] drbd: drbd_req_destroy: don't kref_sub instances from an unrelated object <WDRBD tail recursion re-enable >
Commit: 1d45dcb5f4a713bdba6c99c8939c3efa96b23712 [1d45dcb]
Parents: 7aca1a0ae4
Author: Lars Ellenberg <lars.ellenberg@linbit.com>
Date: 2016년 2월 26일 금요일 오후 9:08:57
Commit Date: 2016년 2월 26일 금요일 오후 9:44:51
drbd: drbd_req_destroy: don't kref_sub instances from an unrelated object

We want to destroy (write) requests in the order we created them
on our resource transfer log.

For each new requests, we get an extra kref
if there is still a preceding WRITE request on the list.

During destruction of a write request, if we find a later write request
on the list, we know "implicitly" that because of the above we have to
drop that extra reference on that request, too.