DW-762 merge linbit patch [2e3a524] drbd: Fix behaviour of detach & attach on a connected primary
Commit: 2e3a524a17fca126410c40341c47e53c25028736 [2e3a524]
Parents: 1379f7703c
Author: Philipp Reisner <philipp.reisner@linbit.com>
Date: 2016년 3월 2일 수요일 오전 12:28:20
Commit Date: 2016년 3월 2일 수요일 오전 12:45:10
drbd: Fix behaviour of detach & attach on a connected primary

..when no write happens between the detach & attach.

Do not consider a node as beeing a CRASHED_PRIMARY if it has an
exposed_data_uuid set or the NEW_CURR_UUID bit set. In both cases
this is a node that survived. It is definitely not a freshly
recovering crashed node.

in disk_states_to_goodness():
Use -2 & +2 in order to make sure that goodness_to_repl_state() modifies
the direction of the resync.
Do not provoke an empty resync if one of the guys is outdated.

In sanitize_state():
If the negotiation ends without a resync make sure to set the disk
state to D_UP_TO_DATE if we have a neighbor that is D_UP_TO_DATE.

In other words, the concept of creating the current uuid late,
creates a lot of corner case code.