DW-1016 [LINBIT] drbd: introduce MDF_PEER_DEVICE_SEEN
커밋: 7cbd0ef5f9e0151390efc4532f5b63293d0d6d24 [7cbd0ef]
상위 항목: 7dfae3cf17
작성자: Lars Ellenberg <lars.ellenberg@linbit.com>
날짜: 2016년 6월 8일 수요일 오전 12:57:55
커밋 날짜: 2016년 6월 8일 수요일 오전 2:14:48
drbd: introduce MDF_PEER_DEVICE_SEEN

A flag in meta data to persist if we have ever seen a certain
peer node with a backing device attached.

If we did not see a backing device, ever, we assume that it
is diskless on purpose (a "DRBD client"), and can ignore it
while calculating the maximum agreeable device size.