DW-1016 [LINBIT] drbd: rework device size handshake (again)
커밋: 4fe4c962f41307c2db2846f18f5137082f339cd9 [4fe4c96]
상위 항목: 75d234c902
작성자: Lars Ellenberg <lars.ellenberg@linbit.com>
날짜: 2016년 6월 8일 수요일 오전 1:30:21
커밋 날짜: 2016년 6월 8일 수요일 오전 2:14:48
drbd: rework device size handshake (again)

Resize was broken in various ways.
I found it hard to split this patch into smaller pieces.

Goal:
  During handshake or during explicitly triggered resize,
  quickly settle on the maximum device size that still fits inside all the
  backend devices of all peers, possibly capped by a user-requested size.

  Do not accidentally shrink a consistent device.
  Do allow explicit shrinking.

  Avoid resizing beyond end of peer backend devices,
  even if those are currently unreachable.

receive_sizes()

  Don't try to fit all information about the size of a peer device
  into a single "max_size" number, but remember all sizes as reported,
  the backend device size if any,
  the user requested size if any,
  and the current device size.
  Store as respective max_size their backend device size,
  possibly capped by the user requested size.

No longer reset to zero this max_size information
just because they detach or disconnect.

get_max_agreeable_size()

  Iterate over all peer devices, minimum-aggregate their respective "max_size"
  (their backend device size possibly capped by their requested-by-user size).

  Peers that are not currently connected are excluded from max_size aggregation,
  even if we may still have stored an informational "max_size" last time they
  talked to us.

  Peers that have never had a backend device, are ignored.

  Report if we currently can see all peer devices we know something about.

If we can NOT currently see all peer devices, we have two options:
we can fall back to the size we last-agreed upon,
or we can "assume-peer-has-space".

If we CAN see all peer devices, and the new "agreed" size is different
from our current size, reset our current size to the new size,
and sent this new information out to every peer we think need to know.

On a (currently) diskless node,
we only accept whatever our peers tell us (but see comments in code).

If the device was extended, we then may still need to resync that new space,
across all peers.  To do so efficiently, we create a "resync-chain",
in the sequence of the node_ids, and trigger a resync to the neighbor devices.
Obviously we ignore diskless nodes, they cannot contribute to the resync.

This patch also increases the code section protected by the conf_update mutex
in receive_sizes().  I'm sure that this is NOT the best idea ever,
but it improved things for me by avoiding a number of races.