DW-1016 [LINBIT] drbd: Fix a BUG by no longer add_timer() twice
커밋: 3a3d1e07b822883cc91dcc6fa650fa35ebaf7d95 [3a3d1e0]
상위 항목: 4fe4c962f4
작성자: Philipp Reisner <philipp.reisner@linbit.com>
날짜: 2016년 6월 8일 수요일 오후 7:03:05
커밋 날짜: 2016년 6월 8일 수요일 오후 8:51:45
drbd: Fix a BUG by no longer add_timer() twice

What was happening quite infrequent:

in drbd_connect() we arm the connect_timer. Then assume a down command comes
in and we go into conn_disconnect(). Which deletes the timer and flushes
the sender workqueue.

Let's assume the timer was able to trigger and queued the connect_work()
worker callback function. When that is executed during the flush
sender workqueue it tries to change state to CONNECTED and probably
fails with SS_CONCURRENT_ST_CHG. I.e. it re-arms the timer!

Later on the user whishes to re-establish the connection. Boom the
timer was already added!

The fix is obviously, when the work function finds the cstate in
something else than C_CONNECTING it drops the connection reference
and exits immediately.