DW-1016 [LINBIT] drbd: Fix how a D_OUTDATED node gets D_UP_TO_DATE after loosing the primary
커밋: bbb3290b1d5a3932884916478c77bd4a4f901f8e [bbb3290]
상위 항목: 7aa8cb9ae2
작성자: Philipp Reisner <philipp.reisner@linbit.com>
날짜: 2016년 6월 13일 월요일 오후 11:55:34
커밋 날짜: 2016년 6월 14일 화요일 오전 12:15:03
drbd: Fix how a D_OUTDATED node gets D_UP_TO_DATE after loosing the primary

In a setup with 3 nodes forming a chain: A - B - C
consider A is R_PRIMARY and D_DISKLESS -> C is D_OUTDATED.
When A goes away ungracefully B becomes temporarily D_CONSISTENT and
goes back to D_UP_TO_DATE as soon as it knows that there is not
different path to the primary.
C does a resync-after-unstable with B. If there is no resync it
need to take over the D_CONSISTENT disk state, which will go over
to D_UP_TO_DATE soon.