DW-1014 fix no resync in spite of existing out-of-sync.
there's no resync if peer reconnects and has same uuid, in spite of existing out-of-sync.
the only thing may affect is disk state(specifically D_INCONSISTENT) when the result of uuid_compare is 0.

added couple of more states which can trigger sync in case of no resync and out-of-sync exists.