DW-1080 fix drbd stops when handshaking
since DW-1014 modification is added, drbd starts sync even though the result of uuid compare is 0.

It operates by checking existing out-of-sync, but sometimes out-of-sync could be inconsistent between two node.

above causes inconsistency of uuid compare result, one of node changes it's repl state to WFBitmapX to try to start sync while another one doesn't.

for coincidence of out-of-sync, we need to also see peer device dirty bit which received with uuid.