DW-1084 [LINBIT-UTILS] drbdmeta: fix device UUID generation
If drbdmeta is called in verbose mode (e.g., drbdadm create-md foo -v),
we read back the command string ("drbdmeta 0 v09..."), instead of the
stdout of the command. As this starts with 'd', and we convert it with
strtol in base 16, we get back a valid number, but a wrong result.
Disable verbose mode while we call "read-dev-uuid".