DW-1084 [LINBIT-UTILS] v9: new resource option max-io-depth
in-kernel, this is res_opts.nr_requests,
in analogy to request_queue.nr_requests.

It serves as an upper bound on the number of in-flight requests we
accept per volume and IO direction.
If we don't limit the submit path, "misbehaved" users may queue up
requests until memory or other system resources are exhausted.