DW-1108 [LINBIT] drbd: A diskless node should pass the u_size along
An example that was wrong before. Nodes:

A - B - C

whre A and C have backing disk(s) and B is diskless.
Have A and C with different size, then issue a resize on A
with a usize slightly smaller than the sizes of A and C.

The result is an endless loop B's size jumping back and
forth between the constraints set by A and C.