DW-1108 [LINBIT] drbd: Do not retry to get the cstate connected when it dropped in the meantime
The bug was a bit hard to spot in the logfiles. We got the connect timer
armed, it fires, schedules the work, starts the change of the cstate. Then
a network failure strikes.

change_cstate() returns with the maybe missleading SS_CONCURRENT_ST_CHG,
but when the cstate is no longer connecting it is a stupid mistake to
try to get connected in 200ms.

  It could happen that the cstate was in C_CONNECTING *again* after
  200ms, and then things started to get confused.