DW-1121 hold I/O when the node is WFBitmapS for peer.
the node sends and receives bitmap when the replication state is WFBitmapS/T, and uses bitmap to calculate how many sectors should be synced.

after peer goes SyncTarget, it still receives out-of-sync newly set on SyncSource.

it possibly causes stopping sync because new out-of-sync could be set on the sector which bm_resync_fo has been already swept.