DW-1141 [LINBIT] drbd: Avoid "initial packet S crossed"
With the old implementation when the two TCP connections get established about
the same time, then both sides send into the TCP connection that they want
to have a DATA_STREAM. Then the whole procedure needs to be aborted and
retried, "initial packet S crossed" gets printed to the log.
It works only when one TCP connection gets established, and without time-
overlap the second one afterwards.

A much smarter approach is to compare the IP addresses of the two endpoints.
One sends DATA_STREAM for the first established TCP connection, the other
one wants to have a CONTROL_STREAM for the first established TCP connection.
Then it magically works out quickly!

It is wire compatible with the previous implementation.