DW-1150 [LINBIT] drbd: fix a two-phase-commit ABORT endless loop
The issue happened when a transaction was aborted by one participant with
RETRY, and then the connection to that participant is lost. When trying to
re-establishing the connection the agreed_protocol is set to 80.

When the __cluster_wide_request() function is used to send out the PREPARE
packets for the next retry, the connection with protocol 80 was skipped,
the old code failed to remove the TWOPC_PREPARED and the TWOPC_RETRY bit!

Fix that by clearing TWOPC_PREPARED for all connections first!