DW-1176 restart drbd receiver when it can't re-try to connect(edited mistyped issue key).
drbd tries to re-connect when it gets failed while connected and receiver is not running.
above causes next connection be disconnected, and it tries to connect again, so peer node does.

some kind of connecting options might be ignored since the node wanders between connection and disconnection.
(discard-my-data is ignored on issue 'DW-1176')