DW-1306 fix bug that resync after unstable isn't triggered.
1. crashed primary flag conflicted with resync after unstable.
=> when primary crashes and re-connected with nodes. it would be getting synced from one node. after it finishes, it gets stable and sends uuid with got stable flag.
crashed primary flag is cleared in w_after_state_change(), we do smt on the peer depends on the connection order.
assume that we send uuid with got stable flag to node 1 first, and then we found finished resync for crashed primary for node 2. node1 does think that I'm still crashed primary.

2. got stable flag is cleared before triggering resync after unstable.
=> one of node gets stable and it sends uuid with got stable flag to start resync. the other nodes receives uuid and set got stable flag to peer_device->uuid_flags
the change of state is detected and it clears got stable flag from peer_device->uuid_flags in the meantime, which races with above.