DW-1382 [LINBIT] drbd: Fix two-phase-commit when we get the prepare packet multiple times
With 913e34a it was fixed for sigma shaped clusters, but it broke
it for plain circular structured clusters.

When we get a prepare from a not we can deduce from that that he
has prepared this transcation as well. If that was is the last
missing TWOPC_YES, set that and send all parents (including the
new one) our agreement.