DW-1382 [LINBIT] drbd: Fix concurrent prepares in two-phase-commit
When a prepare packet arrives at a node concurrently from two peers,
and the node is currently executing an other transaction the following
was possible:

One receiver thread finds the currently ongoing transaction and aborts
it. The other receiver thread quickly grabs the transaction lock and
starts the two-pahse-commit.
The first thread times out waiting for the transaction lock and assumes
it failed to abort the local transaction.

Fixed that by:

If the local transaction changes to that transaction we actually want
to execute, simple add my origin node to the list of parents...