DW-1382 [LINBIT] drbd: Fix abort of a transaction that aborts the currently running transaction
If process_twopc() is called with a P_TWOPC_PREPARE from the worker,
it might need to sleep while it aborts the currently running transaction.

If in the mean time an P_TWOPC_ABORT comes in it does not find the
starting transaction any longer.

Fix that by keeping the starting_tqueued_twopc pointer longer.