DW-1431 [LINBIT] drbd: Make sure new current-uuid gets written into meta-data immediately
This fix cures a 5 seconds window in which we easily get into a split
brain situation.
Example:
Consider 3 nodes, where A is diskless

    A
   / \
  B---C

When A becomes primary an generates a new current-UUID(2), it sends it out
to B and C. Due to the bug B and C do not write it for 5 seconds. Within
the 5 second window C is rebootet, A generates the next current-UUID(3)
sind it lost C.
When C comes back and wants to resync with B it is still on UUID(1).
-> We see a split brain.