DW-1293 move acquiring and releasing att_mod_mutex
when node is about to go primary, it does not set volume attribute and needs to mount volume.

deadlock can be occured when writing data on md is necessary while mounting.

acquiring and releasing att_mod_mutex are needed only when setting volume attribute to read-only, which means no write is needed during mount.