DW-1569 [LINBIT-UTILS] Fix creating peer-device objects
It was incomplete. It did it only for one host of the
resource (which might be one that has nothing to do
with this connection).

It needs to do it for the hosts of the connection.

The bug was discovered because commit 326e6e9
introduces a loop that goes over the peer devices of a volume.
With the old code it was easy to hit a volume without peer_devices...