DW-1569 [LINBIT-UTILS] drbdadm: send SIGKILL to child if parent exits
If the parent (i.e. drbdadm) fork/execs a child, set up the child in a
way that it receives a SIGKILL if the parent exits. This is for example
useful when a SLEEPS_ timeout is reached and drbdadm exits.

We only want that behavior if the command is executed from drbdadm, but
not in the command itself (e.g., a directly called "drbdsetup events2"
should continue if the shell dies).

This mechanism requires that drbdadm directly calls the correct binary
(e.g., drbdsetup-84). If the general drbdsetup is called and then
execs for example drbdsetup-84, PR_SET_PDEATHSIG is not set for the
grand child. The value is preserved across execve(2), but cleared on
fork(2). drbdadm has to directly call the final binary.