DW-1569 [LINBIT-UTILS] drbd udev: fix inconsistent inheritance of implicit volumes
The "implicit-ness" of volume definitions on the resource level
was not inherited by the host sections.

That confused the generation of udev symlinks to sometimes
(but not always) inconsistently use "/0" in udev symlinks
even for implicit volumes.

Also, such volume definitions would forget to print the "disk"
definition in a non-verbose dump.

This patch fixes that, and fixes the udev symlinks so implicit volumes
would behave as intended and never add the "/0".

For those that "accidentally" had these unintended symlinks, or want
them for any reason, it adds a new global option udev-always-use-vnr,
which is disabled by default, but enabled in the example global_common.conf.