DW-1569 [LINBIT-UTILS] v9 adjust: fix segfault in post_parse of second drbdsetup show
You can call "adjust" for "all" resources,
or for an explicit list of resources, which typically is exactly one resources.

drbdadm adjust all                   # ok
drbdadm adjust exactly-this-resource # ok
drbdadm adjust res1 res2 ...         # NOT OK

If called for all resources, we call "drbdsetup show", once,
for the current "live" configuration of all resources.

If called for a list of explicit resources,
we called "drbdsetup show that-resource",
for each resource in turn.

When "adjust" was called with more than one explicitly listed resource,
it hit a segfault in post_parse of the second drbdsetup show.

Avoid that by checking for "adjust with more than one resource" first,
and if so, call "drbdsetup show" only once, like in the "all" case.