DW-1569 [LINBIT-UTILS] drbdsetup,v9: fix cstate output
So far this searched in a list of all connections for all resources and
then has happy if it found a matching peer. This is not enough as it
could find a matching peer for an unrelated connection.

Fix that by narrowing the possible connections to the one resource we
are actually looking for.