DW-1469 [LINBIT] drbd: Resync after online grow
In case a primary node is present, resync from the primary to the neighbours
of the primaries. The secondaries farther away do start a resync.

In case no primary node is present, do resyncs originating the initator
node of the transaction. They follow exactly the flow of the two-pc
packets.